refactor: removed visibility from the link
