+++ src/components/custom/_signup-public.tsx
@@ -5,7 +5,7 @@ import AgreementLabel from './_agreement-label'
 import { Input, Button } from 'components/form'
 import { Header, LinkText, QueryImage, Text } from 'components/elements'
 import { localize } from 'components/localization'
-import { Flex, Box, Container } from 'components/containers'
+import { Flex, Box, Container, Desktop, Mobile } from 'components/containers'
 import { deriv_app_url } from 'common/constants'
 import { useCountryRule } from 'components/hooks/use-country-rule'
 import device from 'themes/device'
@@ -14,7 +14,6 @@ import Apple from 'images/svg/custom/apple-40.svg'
 import Facebook from 'images/svg/custom/facebook-40.svg'
 import Google from 'images/svg/custom/google-40.svg'
 import Arrow from 'images/svg/custom/chevron-right.svg'
-import { Desktop, Mobile } from 'components/containers/visibility'
 
 type SignupPublicProps = {
     autofocus?: boolean
