+++ src/pages/payment-methods/_dp2p.tsx
@@ -1,10 +1,9 @@
 import React, { Fragment, useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
-import { Desktop, Mobile } from 'components/containers/visibility'
 import device, { size } from 'themes/device'
 import { Text, Header, QueryImage } from 'components/elements'
-import { Flex } from 'components/containers'
+import { Flex, Desktop, Mobile } from 'components/containers'
 import { Button, LinkButton } from 'components/form'
 import { localize, Localize } from 'components/localization'
 import { deriv_dp2p_app_url, dp2p_google_play_url } from 'common/constants'
