+++ src/pages/trader-tools/pip-calculator/_pip-calculator.tsx
@@ -49,9 +49,8 @@ import {
 } from 'components/elements'
 import Input from 'components/form/input'
 import RightArrow from 'images/svg/tools/black-right-arrow.svg'
-import { Flex } from 'components/containers'
+import { Flex, Desktop, Mobile } from 'components/containers'
 import { localize, Localize } from 'components/localization'
-import { Desktop, Mobile } from 'components/containers/visibility'
 
 const PipCalculator = () => {
     const query = graphql`
