+++ src/pages/trader-tools/pnl-for-multipliers/_pnl_multipliers_calculator.tsx
@@ -35,7 +35,6 @@ import {
     StyledSection,
     SwapTabSelector,
 } from '../common/_style'
-import { Desktop, Mobile } from 'components/containers/visibility'
 import { localize, Localize } from 'components/localization'
 import {
     Accordion,
@@ -45,7 +44,7 @@ import {
     QueryImage,
     Text,
 } from 'components/elements'
-import { Flex } from 'components/containers'
+import { Flex, Desktop, Mobile } from 'components/containers'
 import Input from 'components/form/input'
 import RightArrow from 'images/svg/tools/black-right-arrow.svg'
